requirements.txt
sqlmodel~=0.0.24
SQLAlchemy~=2.0.40
fastapi~=0.115.12
pydantic~=2.11.4
starlette~=0.46.2
pydantic-settings~=2.9.1
nanoid~=2.0.0
alembic~=1.15.2
pytest~=8.3.5
SQLAlchemy-Utils~=0.41.2
python-dotenv~=1.1.0